+++ src/main/java/com/db/phase4/dao/RentalDao.java
@@ -1,5 +1,6 @@
-package com.db.phase4.rental;
+package com.db.phase4.dao;
 
+import com.db.phase4.dto.RentalDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
