+++ src/main/java/com/db/phase4/service/RentalService.java
@@ -1,5 +1,7 @@
-package com.db.phase4.rental;
+package com.db.phase4.service;
 
+import com.db.phase4.dao.RentalDao;
+import com.db.phase4.dto.RentalDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
